Widen payload_s dport from uint16_t to int32_t for -1 convention
This change extends Jack's existing -1 convention (already used for
sport to mean "any source port") to dport. This allows payload modules
to declare themselves as default payloads via metadata (dport=-1)
rather than calling add_default_payload() directly from init_module().

The report_s struct already uses int32_t for dport with the same
semantics ("/* -1 for all */"), so this change also adds consistency
within the mod_entry_t union.

All existing payload modules continue to work unchanged as they set
positive port values (53, 80, 443, etc.).

+++ src/unilib/modules.h
@@ -65,7 +65,7 @@ typedef struct mod_entry_t {
 		struct payload_mod {
 			int16_t proto;
 			int32_t sport;
-			uint16_t dport;
+			int32_t dport;	/* -1 for default payload (matches sport convention) */
 			uint16_t payload_group;
 		} payload_s;
 		struct report_mod {
